+++ src/main/java/com/commercetools/sync/producttypes/ProductTypeSyncOptions.java
@@ -20,7 +20,6 @@ public final class ProductTypeSyncOptions extends BaseSyncOptions<ProductType, P
         @Nullable final BiConsumer<String, Throwable> updateActionErrorCallBack,
         @Nullable final Consumer<String> updateActionWarningCallBack,
         final int batchSize,
-        final boolean allowUuid,
         @Nullable final TriFunction<List<UpdateAction<ProductType>>, ProductTypeDraft, ProductType,
                 List<UpdateAction<ProductType>>> beforeUpdateCallback,
         @Nullable final Function<ProductTypeDraft, ProductTypeDraft> beforeCreateCallback
@@ -31,7 +30,6 @@ public final class ProductTypeSyncOptions extends BaseSyncOptions<ProductType, P
             updateActionErrorCallBack,
             updateActionWarningCallBack,
             batchSize,
-            allowUuid,
             beforeUpdateCallback,
             beforeCreateCallback
         );